Slowly changing Date into String

+++ YlperJ-core/src/main/java/com/shoter/ylper/core/Bookings/BookingRepository.java
@@ -15,5 +15,5 @@ public interface BookingRepository extends Repository<Booking> {
     boolean bookingExistsInGivenTime(long carId, Date time);
     boolean bookingExistsInGivenTime(long carId, Date startTime, Date endTime);
 
-    List<FindCarResult> findProperCar(Date startTime, Date endTime, Integer carLuxuryCategoryId, List<Integer> carFeatureIds, Point searchLocation);
+    List<FindCarResult> findProperCar(Date startTime, Date endTime, byte carLuxuryCategoryId, List<Integer> carFeatureIds, Point searchLocation);
 }
