+++ YlperJ-core/src/main/java/com/shoter/ylper/core/Bookings/BookingService.java
@@ -20,5 +20,5 @@ public interface BookingService {
     boolean bookingExistsInGivenTimeForCar(long carId, Date time);
     boolean bookingExistsInGivenTimeForCar(long carId, Date startTime, Date endTime);
 
-    List<FindCarResult> findProperCar(Date startTime, Date endTime, int carLuxuryCategoryId, List<Integer> carFeatureIds, Point searchLocation);
+    List<FindCarResult> findProperCar(Date startTime, Date endTime, byte carLuxuryCategoryId, List<Integer> carFeatureIds, Point searchLocation);
 }
